+++ test/config_cmd-test.js
@@ -1,8 +1,8 @@
 /* global describe it afterEach */
 
 import sinon from 'sinon'
-import * as config from './config'
-import configCmd from './config_cmd'
+import * as config from '../src//config'
+import configCmd from '../src/config_cmd'
 
 const sandbox = sinon.sandbox.create()
