+++ test/link-test.js
@@ -2,8 +2,8 @@
 
 import assert from 'assert'
 import sinon from 'sinon'
-import * as config from './config'
-import * as link from './link'
+import * as config from '../src/config'
+import * as link from '../src/link'
 import fs from 'fs'
 
 const sandbox = sinon.sandbox.create()
